requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn==1.3.2
xgboost==2.0.3
scipy
tslearn
joblib==1.3.2
loky==3.4.0
plotly>=5.10
streamlit-aggrid==0.3.5
requests>=2.28.0

# Google Cloud
google-api-python-client
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-storage
db-dtypes
pyarrow 